# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Cicsa Api",
        "description": "Cicsa Controller",
        "version": "1"
    },
    "paths": {
        "/api/bitacoras": {
            "get": {
                "operationId": "c08ec686f0bb978291e1453510020270",
                "responses": {
                    "200": {
                        "description": "Muestra una lista de bitacoras."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/bitacoras/{id}": {
            "get": {
                "operationId": "f192ed65689f0c76db178c033173ff07",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la bitácora",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Muestra una bitácora específica."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "operationId": "4ee01ccb80a47369dc60482f5351f974",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la bitácora",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Actualiza una bitácora."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}